refactor: update import paths for SearchBar and Movie interfaces, and migrate Carousel component to core directory, enhancing code organization and maintainability
